+++ src/main/java/com/donkeys_today/server/application/user/UserService.java
@@ -4,24 +4,21 @@ import com.donkeys_today.server.application.auth.JwtUtil;
 import com.donkeys_today.server.application.user.event.UserSignUpEvent;
 import com.donkeys_today.server.common.constants.Constants;
 import com.donkeys_today.server.domain.user.User;
+import com.donkeys_today.server.presentation.auth.dto.request.UserSignInRequest;
+import com.donkeys_today.server.presentation.auth.dto.request.UserSignUpRequest;
 import com.donkeys_today.server.presentation.auth.dto.response.TokenReissueResponse;
 import com.donkeys_today.server.presentation.user.dto.requset.UserNamePatchRequest;
-import com.donkeys_today.server.presentation.user.dto.requset.UserSignInRequest;
-import com.donkeys_today.server.presentation.user.dto.requset.UserSignUpRequest;
 import com.donkeys_today.server.presentation.user.dto.response.UserDeleteResponse;
 import com.donkeys_today.server.presentation.user.dto.response.UserInfoResponse;
 import com.donkeys_today.server.presentation.user.dto.response.UserNamePatchResponse;
-import com.donkeys_today.server.presentation.user.dto.response.UserSignInResponse;
-import com.donkeys_today.server.presentation.user.dto.response.UserSignUpResponse;
-import com.donkeys_today.server.support.dto.type.ErrorType;
-import com.donkeys_today.server.support.exception.UnauthorizedException;
+import com.donkeys_today.server.presentation.auth.dto.response.UserSignInResponse;
+import com.donkeys_today.server.presentation.auth.dto.response.UserSignUpResponse;
 import com.donkeys_today.server.support.jwt.JwtProvider;
 import com.donkeys_today.server.support.jwt.Token;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import org.springframework.util.StringUtils;
 
 @Service
 @RequiredArgsConstructor
