+++ src/main/java/com/donkeys_today/server/application/user/sterategy/KakaoAuthStrategy.java
@@ -3,8 +3,8 @@ package com.donkeys_today.server.application.user.sterategy;
 import com.donkeys_today.server.domain.user.Platform;
 import com.donkeys_today.server.domain.user.User;
 import com.donkeys_today.server.infrastructure.user.UserRepository;
-import com.donkeys_today.server.presentation.user.dto.requset.UserSignInRequest;
-import com.donkeys_today.server.presentation.user.dto.requset.UserSignUpRequest;
+import com.donkeys_today.server.presentation.auth.dto.request.UserSignInRequest;
+import com.donkeys_today.server.presentation.auth.dto.request.UserSignUpRequest;
 import com.donkeys_today.server.support.dto.type.ErrorType;
 import com.donkeys_today.server.support.exception.auth.SignInException;
 import com.donkeys_today.server.support.exception.auth.SignUpException;
