+++ src/main/java/com/donkeys_today/server/application/user/sterategy/SocialRegisterSterategy.java
@@ -1,8 +1,8 @@
 package com.donkeys_today.server.application.user.sterategy;
 
 import com.donkeys_today.server.domain.user.User;
-import com.donkeys_today.server.presentation.user.dto.requset.UserSignInRequest;
-import com.donkeys_today.server.presentation.user.dto.requset.UserSignUpRequest;
+import com.donkeys_today.server.presentation.auth.dto.request.UserSignInRequest;
+import com.donkeys_today.server.presentation.auth.dto.request.UserSignUpRequest;
 
 public interface SocialRegisterSterategy {
