+++ src/main/java/com/donkeys_today/server/presentation/auth/dto/request/UserSignInRequest.java
@@ -1,4 +1,4 @@
-package com.donkeys_today.server.presentation.user.dto.requset;
+package com.donkeys_today.server.presentation.auth.dto.request;
 
 import io.swagger.v3.oas.annotations.media.Schema;
